startup script directory permissions cleanup
adding some additional checks to only apply directory permissions
checks if they're present. there are slightly different start
conditions based on if the container is being run for the first
time versus being restarted.

+++ assets/startup.sh
@@ -3,13 +3,18 @@
 DEFAULT_NTP="time.cloudflare.com"
 CHRONY_CONF_FILE="/etc/chrony/chrony.conf"
 
-# update permissions on chrony directories
-mkdir -p /run/chrony
-chown -R chrony:chrony /run/chrony /var/lib/chrony
-chmod o-rx /run/chrony
+# confirm correct permissions on chrony run directory
+if [ -d /run/chrony ]; then
+  chown -R chrony:chrony /run/chrony
+  chmod o-rx /run/chrony
+  # remove previous pid file if it exist
+  rm -f /var/run/chrony/chronyd.pid
+fi
 
-# remove previous pid file if it exist
-rm -f /var/run/chrony/chronyd.pid
+# confirm correct permissions on chrony variable state directory
+if [ -d /var/lib/chrony ]; then
+  chown -R chrony:chrony /var/lib/chrony
+fi
 
 ## dynamically populate chrony config file.
 {
